feat(frontend): remove temp file
