+++ frontend/src/lib/components/apps/editor/SubGridEditor.svelte
@@ -1,8 +1,97 @@
 <script lang="ts">
-	import type { GridItem } from '../types'
-	import Test from './Test.svelte'
+	import { classNames } from '$lib/utils'
+	import { getContext } from 'svelte'
+	import Grid from 'svelte-grid'
+	import { columnConfiguration, isFixed, toggleFixed } from '../gridUtils'
+	import type { AppEditorContext, GridItem } from '../types'
+	import Component from './Component.svelte'
 
 	export let subGrid: GridItem[]
+
+	const { app, connectingInput, selectedComponent } =
+		getContext<AppEditorContext>('AppEditorContext')
+
+	let pointerdown = false
+	const onpointerdown = () => {
+		pointerdown = true
+	}
+	const onpointerup = () => {
+		pointerdown = false
+	}
+
+	function removeGridElement(data: any): void {
+		throw new Error('Function not implemented.')
+	}
+
+	function selectComponent(id: string) {
+		if (!$connectingInput.opened) {
+			$selectedComponent = id
+		}
+	}
+
+	function lock(gridComponent: GridItem) {
+		let fComponent = $app.grid.find((c) => c.id === gridComponent.id)
+		if (fComponent) {
+			fComponent = toggleFixed(fComponent)
+		}
+	}
 </script>
 
-<Test bind:lazyGrid={subGrid} />
+<div class="pb-2 relative w-full z-20">
+	<div
+		class="px-4 pt-4 overflow-auto {$connectingInput?.opened ? '' : ''}"
+		on:pointerdown={onpointerdown}
+		on:pointerleave={onpointerup}
+		on:pointerup={onpointerup}
+	>
+		<Grid
+			bind:items={subGrid}
+			let:dataItem
+			rowHeight={36}
+			cols={columnConfiguration}
+			fastStart={true}
+			gap={[4, 2]}
+		>
+			{#each subGrid as gridComponent (gridComponent.id)}
+				{#if gridComponent.data.id === dataItem.data.id}
+					<!-- svelte-ignore a11y-click-events-have-key-events -->
+					{#if $connectingInput.opened}
+						{#if $connectingInput.opened}
+							<div
+								on:pointerenter={() => ($connectingInput.hoveredComponent = gridComponent.data.id)}
+								on:pointerleave={() => ($connectingInput.hoveredComponent = undefined)}
+								class="absolute w-full h-full bg-black border-2 bg-opacity-25 z-20 flex justify-center items-center"
+							/>
+						{/if}
+						<div
+							style="transform: translate(-50%, -50%);"
+							class="absolute w-fit justify-center bg-indigo-500/90 left-[50%] top-[50%] z-50 px-6 rounded border text-white py-2 text-5xl center-center"
+						>
+							{dataItem.data.id}
+						</div>
+					{/if}
+
+					<!-- svelte-ignore a11y-click-events-have-key-events -->
+					<div
+						on:pointerdown={() => selectComponent(dataItem.data.id)}
+						class={classNames(
+							'h-full w-full center-center',
+							$selectedComponent === dataItem.data.id ? 'active-grid-item' : '',
+							gridComponent.data.card ? 'border border-gray-100' : '',
+							'top-0'
+						)}
+					>
+						<Component
+							{pointerdown}
+							bind:component={gridComponent.data}
+							selected={$selectedComponent === dataItem.data.id}
+							locked={isFixed(gridComponent)}
+							on:delete={() => removeGridElement(gridComponent.data)}
+							on:lock={() => lock(gridComponent)}
+						/>
+					</div>
+				{/if}
+			{/each}
+		</Grid>
+	</div>
+</div>
